-improve  minor details(renaming, delete commented code,rearrange classes by packages)
