--- database/src/main/java/by/itechart/demo/post/repository/ElasticPostRepository.java
@@ -1,6 +1,0 @@
-package by.itechart.demo.post.repository;
-
-//@Repository
-//public interface ElasticPostRepository extends ElasticsearchRepository<Post,Long> {
-//
-//}
