+++ database/src/main/java/by/itechart/demo/user/model/Role.java
@@ -1,5 +1,5 @@
 package by.itechart.demo.user.model;
 
 public enum Role {
-    Admin,User
+    Admin,User,Anonym
 }
