+++ database/src/main/java/by/itechart/demo/user/model/User.java
@@ -22,6 +22,8 @@ public class User {
     @Column(name = "username")
     private String userName;
     private String password;
-    private String role;
+
+    @Enumerated(EnumType.STRING)
+    private Role role;
 
 }
