+++ database/src/main/resources/application.yml
@@ -1,4 +1,10 @@
 spring:
+  profiles:
+    active: "dev"
+---
+
+spring:
+  profiles: dev
   datasource:
      driver-class-name: org.postgresql.Driver
      url: jdbc:postgresql://localhost:5432/postgres
@@ -38,3 +44,5 @@ logging:
 jwt:
   secret: demo
   validity: 36000
+
+
